Fix local task replay
Previously, local tasks were replayed incorrectly. The last
transformation in the task was being run in a new JVM, but all other
transformations were being run in the same JVM as the
EventLogReader. If one of the earlier transformations threw an
exception, that exception would occur in the EventLogReader instead of
in the new JVM.

Instead, this commit replays tasks by first ensuring that the shuffle
outputs from all parent stages have been computed, and then executing
the entire task in a separate JVM that acts like a slave.